+++ Dockerfile
@@ -8,7 +8,7 @@ RUN pip install --upgrade pip
 
 RUN pip install -r requirements.txt
 
-COPY webapp/* .
+COPY . .
 
 ENTRYPOINT [ "python" ]
