+++ hur_git.txt
@@ -1,0 +1,2 @@
+1. git status
+2. git add [filnamn eller .]
