+++ templates/about.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<title>About Us</title>
+	<link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
+</head>
+<body>
+	<h1>About Us</h1>
+	<p>We are a team of industrial economics and managment studnets, dedicated to creating high-quality SensorSytem for seniors.</p>
+	<h2>Our Mission</h2>
+	<p>Our mission is to help seiors to stay at home for a longer time, while being well kwon about their health condition.</p>
+	<h2>Our Team</h2>
+	<ul>
+		<li>Salima - Scrum Master</li>
+        <li>Hamode - Businessanalyst</li>
+		<li>Qlirim - UI/UX Designer</li>
+		<li>Shams - Sekreterare</li>
+        <li>Fatime - Testansvarig</li>
+	</ul>
+	<audio controls autoplay tabindex="0">
+		<source src="{{ url_for('static', filename='song.mp3') }}" type="audio/mpeg">
+		Your browser does not support the audio element.
+	</audio>
+</body>
+</html>
