+++ templates/home.html
@@ -1,0 +1,47 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Home Page</title>
+    <style>
+        body {
+            background-image: url("/static/myimage.jpg");
+            background-size: cover;
+        }
+        .header {
+            text-align: center;
+            font-size: 3rem;
+            margin-top: 5rem;
+            color: white;
+            text-shadow: 2px 2px 4px #000000;
+        }
+        .button-container {
+            display: flex;
+            justify-content: center;
+            margin-top: 2rem;
+        }
+        .button {
+            background-color: #008CBA;
+            color: white;
+            font-size: 1.5rem;
+            padding: 1rem 3rem;
+            border-radius: 5px;
+            text-decoration: none;
+            margin: 1rem;
+            box-shadow: 2px 2px 4px #000000;
+        }
+    </style>
+</head>
+<body>
+    <div class="header">
+        <h1>Welcome to SensorSytem</h1>
+        <h4>Please Signup or Login if you already have an account:</h4>
+    </div>
+    <div class="button-container">
+        <a href="{{ url_for('login') }}" class="button">Login</a>
+        <a href="{{ url_for('signup') }}" class="button">Signup</a>
+        <a href="{{ url_for('about') }}" class="button">About Us</a>
+    </div>
+</body>
+</html>
+
+
