--- webapp/templates/base.html
@@ -1,8 +1,0 @@
-<html>
-    <body>
-        {% block content %}
-
-        {% endblock %}
-    </body>
-</html>
-
