--- webapp/templates/index.html
@@ -1,5 +1,0 @@
-{% extens "base.html" %}
-
-{% block content %}
-<h1>Index</h1>
-{% endblock %}
